separated cluster and server into different files
